+++ CHANGELOG.md
@@ -19,7 +19,7 @@ All notable changes to [Earthly](https://github.com/earthly/earthly) will be doc
 - An authentication attempt with an expired auth token will result in a `auth token expired` error instead of `unauthorized`.
 - A successful authentication with an auth token will display a warning with time left before token expires if it's 14 days or under.
 - The command `earthly registry` will attempt to use the selected org if no org is specified.
- 
+- Clarify error messages when failing to pass secrets to a build.
 
 ## v0.7.21 - 2023-10-24